chore: split UserProvider into two: User and Authentication
