+++ apps/meteor/client/providers/UserProvider/UserProvider.tsx
@@ -1,7 +1,7 @@
 import type { IRoom, ISubscription, IUser } from '@rocket.chat/core-typings';
 import { useLocalStorage } from '@rocket.chat/fuselage-hooks';
-import type { LoginService, SubscriptionWithRoom } from '@rocket.chat/ui-contexts';
-import { UserContext, useEndpoint, useSetting } from '@rocket.chat/ui-contexts';
+import type { SubscriptionWithRoom } from '@rocket.chat/ui-contexts';
+import { UserContext, useEndpoint } from '@rocket.chat/ui-contexts';
 import { Meteor } from 'meteor/meteor';
 import type { ContextType, ReactElement, ReactNode } from 'react';
 import React, { useEffect, useMemo } from 'react';
@@ -15,32 +15,12 @@ import { createReactiveSubscriptionFactory } from '../../lib/createReactiveSubsc
 import { useCreateFontStyleElement } from '../../views/account/accessibility/hooks/useCreateFontStyleElement';
 import { useDeleteUser } from './hooks/useDeleteUser';
 import { useEmailVerificationWarning } from './hooks/useEmailVerificationWarning';
-import { useLDAPAndCrowdCollisionWarning } from './hooks/useLDAPAndCrowdCollisionWarning';
 import { useUpdateAvatar } from './hooks/useUpdateAvatar';
 import { useUpdateCustomUserStatus } from './hooks/useUpdateCustomUserStatus';
 
 const getUserId = (): string | null => Meteor.userId();
 
 const getUser = (): IUser | null => Meteor.user() as IUser | null;
-
-const capitalize = (str: string): string => str.charAt(0).toUpperCase() + str.slice(1);
-
-const config: Record<string, Partial<LoginService>> = {
-	'apple': { title: 'Apple', icon: 'apple' },
-	'facebook': { title: 'Facebook', icon: 'facebook' },
-	'twitter': { title: 'Twitter', icon: 'twitter' },
-	'google': { title: 'Google', icon: 'google' },
-	'github': { title: 'Github', icon: 'github' },
-	'github_enterprise': { title: 'Github Enterprise', icon: 'github' },
-	'gitlab': { title: 'Gitlab', icon: 'gitlab' },
-	'dolphin': { title: 'Dolphin', icon: 'dophin' },
-	'drupal': { title: 'Drupal', icon: 'drupal' },
-	'nextcloud': { title: 'Nextcloud', icon: 'nextcloud' },
-	'tokenpass': { title: 'Tokenpass', icon: 'tokenpass' },
-	'meteor-developer': { title: 'Meteor', icon: 'meteor' },
-	'wordpress': { title: 'WordPress', icon: 'wordpress' },
-	'linkedin': { title: 'Linkedin', icon: 'linkedin' },
-};
 
 const logout = (): Promise<void> =>
 	new Promise((resolve, reject) => {
@@ -56,16 +36,11 @@ const logout = (): Promise<void> =>
 		});
 	});
 
-export type LoginMethods = keyof typeof Meteor extends infer T ? (T extends `loginWith${string}` ? T : never) : never;
-
 type UserProviderProps = {
 	children: ReactNode;
 };
 
 const UserProvider = ({ children }: UserProviderProps): ReactElement => {
-	const isLdapEnabled = useSetting<boolean>('LDAP_Enable');
-	const isCrowdEnabled = useSetting<boolean>('CROWD_Enable');
-
 	const userId = useReactiveValue(getUserId);
 	const user = useReactiveValue(getUser);
 	const [userLanguage, setUserLanguage] = useLocalStorage('userLanguage', '');
@@ -76,9 +51,6 @@ const UserProvider = ({ children }: UserProviderProps): ReactElement => {
 	const createFontStyleElement = useCreateFontStyleElement();
 	createFontStyleElement(user?.settings?.preferences?.fontSize);
 
-	const loginMethod: LoginMethods = (isLdapEnabled && 'loginWithLDAP') || (isCrowdEnabled && 'loginWithCrowd') || 'loginWithPassword';
-
-	useLDAPAndCrowdCollisionWarning();
 	useEmailVerificationWarning(user ?? undefined);
 
 	useUpdateCustomUserStatus();
@@ -103,75 +75,9 @@ const UserProvider = ({ children }: UserProviderProps): ReactElement => {
 
 				return ChatRoom.find(query, options).fetch();
 			}),
-			loginWithToken: (token: string): Promise<void> =>
-				new Promise((resolve, reject) =>
-					Meteor.loginWithToken(token, (err) => {
-						if (err) {
-							return reject(err);
-						}
-						resolve(undefined);
-					}),
-				),
-			loginWithPassword: (user: string | { username: string } | { email: string } | { id: string }, password: string): Promise<void> =>
-				new Promise((resolve, reject) => {
-					Meteor[loginMethod](user, password, (error) => {
-						if (error) {
-							reject(error);
-							return;
-						}
-
-						resolve();
-					});
-				}),
 			logout,
-			loginWithService: <T extends LoginService>({ service, clientConfig = {} }: T): (() => Promise<true>) => {
-				const loginMethods = {
-					'meteor-developer': 'MeteorDeveloperAccount',
-				} as const;
-
-				const loginWithService = `loginWith${loginMethods[service] || capitalize(String(service || ''))}`;
-
-				const method: (config: unknown, cb: (error: any) => void) => Promise<true> = (Meteor as any)[loginWithService] as any;
-
-				if (!method) {
-					return () => Promise.reject(new Error('Login method not found'));
-				}
-
-				return () =>
-					new Promise((resolve, reject) => {
-						method(clientConfig, (error: any): void => {
-							if (!error) {
-								resolve(true);
-								return;
-							}
-							reject(error);
-						});
-					});
-			},
-			queryAllServices: createReactiveSubscriptionFactory(() =>
-				ServiceConfiguration.configurations
-					.find(
-						{
-							showButton: { $ne: false },
-						},
-						{
-							sort: {
-								service: 1,
-							},
-						},
-					)
-					.fetch()
-					.map(
-						({ appId: _, ...service }) =>
-							({
-								title: capitalize(String((service as any).service || '')),
-								...service,
-								...(config[(service as any).service] ?? {}),
-							} as any),
-					),
-			),
 		}),
-		[userId, user, loginMethod],
+		[userId, user],
 	);
 
 	useEffect(() => {
