+++ apps/meteor/client/sidebar/Sidebar.stories.tsx
@@ -1,5 +1,5 @@
 import type { ISetting } from '@rocket.chat/core-typings';
-import type { LoginService, SubscriptionWithRoom } from '@rocket.chat/ui-contexts';
+import type { SubscriptionWithRoom } from '@rocket.chat/ui-contexts';
 import { UserContext, SettingsContext } from '@rocket.chat/ui-contexts';
 import type { Meta, Story } from '@storybook/react';
 import type { ObjectId } from 'mongodb';
@@ -98,10 +98,6 @@ const userContextValue: ContextType<typeof UserContext> = {
 	querySubscription: () => [() => () => undefined, () => undefined],
 	queryRoom: () => [() => () => undefined, () => undefined],
 
-	queryAllServices: () => [() => (): void => undefined, (): LoginService[] => []],
-	loginWithService: () => () => Promise.reject('loginWithService not implemented'),
-	loginWithPassword: async () => Promise.reject('loginWithPassword not implemented'),
-	loginWithToken: async () => Promise.reject('loginWithToken not implemented'),
 	logout: () => Promise.resolve(),
 };
