+++ packages/mock-providers/src/MockedUserContext.tsx
@@ -1,4 +1,3 @@
-import type { LoginService } from '@rocket.chat/ui-contexts';
 import { UserContext } from '@rocket.chat/ui-contexts';
 import React from 'react';
 import type { ContextType } from 'react';
@@ -23,10 +22,6 @@ const userContextValue: ContextType<typeof UserContext> = {
 	querySubscription: () => [() => () => undefined, () => undefined],
 	queryRoom: () => [() => () => undefined, () => undefined],
 
-	queryAllServices: () => [() => (): void => undefined, (): LoginService[] => []],
-	loginWithService: () => () => Promise.reject('loginWithService not implemented'),
-	loginWithPassword: async () => Promise.reject('loginWithPassword not implemented'),
-	loginWithToken: async () => Promise.reject('loginWithToken not implemented'),
 	logout: () => Promise.resolve(),
 };
