+++ packages/ui-contexts/src/UserContext.ts
@@ -25,16 +25,6 @@ export type FindOptions<TSchema extends Document = Document> = {
 	sort?: Sort<TSchema>;
 };
 
-export type LoginService = {
-	clientConfig: unknown;
-
-	title: string;
-	service: 'meteor-developer';
-
-	buttonLabelText?: string;
-	icon?: string;
-};
-
 export type UserContextValue = {
 	userId: string | null;
 	user: IUser | null;
@@ -56,13 +46,7 @@ export type UserContextValue = {
 		query: SubscriptionQuery,
 		options?: FindOptions,
 	) => [subscribe: (onStoreChange: () => void) => () => void, getSnapshot: () => SubscriptionWithRoom[]];
-
-	loginWithPassword: (user: string | { username: string } | { email: string } | { id: string }, password: string) => Promise<void>;
-	loginWithToken: (user: string) => Promise<void>;
 	logout: () => Promise<void>;
-
-	queryAllServices(): [subscribe: (onStoreChange: () => void) => () => void, getSnapshot: () => LoginService[]];
-	loginWithService<T extends LoginService>(service: T): () => Promise<true>;
 };
 
 export const UserContext = createContext<UserContextValue>({
@@ -72,10 +56,5 @@ export const UserContext = createContext<UserContextValue>({
 	querySubscription: () => [() => (): void => undefined, (): undefined => undefined],
 	queryRoom: () => [() => (): void => undefined, (): undefined => undefined],
 	querySubscriptions: () => [() => (): void => undefined, (): [] => []],
-
-	queryAllServices: () => [() => (): void => undefined, (): LoginService[] => []],
-	loginWithService: () => () => Promise.reject('loginWithService not implemented'),
-	loginWithPassword: async () => Promise.reject('loginWithPassword not implemented'),
-	loginWithToken: async () => Promise.reject('loginWithToken not implemented'),
 	logout: () => Promise.resolve(),
 });
