+++ packages/ui-contexts/src/hooks/useLoginWithPassword.ts
@@ -1,8 +1,8 @@
 import { useContext } from 'react';
 
-import { UserContext } from '../UserContext';
+import { AuthenticationContext } from '../AuthenticationContext';
 
 export const useLoginWithPassword = (): ((
 	user: string | { username: string } | { email: string } | { id: string },
 	password: string,
-) => Promise<void>) => useContext(UserContext).loginWithPassword;
+) => Promise<void>) => useContext(AuthenticationContext).loginWithPassword;
