+++ packages/ui-contexts/src/hooks/useLoginWithService.ts
@@ -1,10 +1,10 @@
 import { useContext, useMemo } from 'react';
 
-import type { LoginService } from '../UserContext';
-import { UserContext } from '../UserContext';
+import type { LoginService } from '../AuthenticationContext';
+import { AuthenticationContext } from '../AuthenticationContext';
 
 export const useLoginWithService = <T extends LoginService>(service: T): (() => Promise<true>) => {
-	const { loginWithService } = useContext(UserContext);
+	const { loginWithService } = useContext(AuthenticationContext);
 
 	return useMemo(() => {
 		return loginWithService(service);
