+++ packages/ui-contexts/src/hooks/useLoginWithToken.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
 
-import { UserContext } from '../UserContext';
+import { AuthenticationContext } from '../AuthenticationContext';
 
-export const useLoginWithToken = (): ((token: string) => Promise<void>) => useContext(UserContext).loginWithToken;
+export const useLoginWithToken = (): ((token: string) => Promise<void>) => useContext(AuthenticationContext).loginWithToken;
